Highlight app or process with high resource utilization
When an app or process is utilizing a large amount of system resources, provide
a visual indication of that fact.

The implementation calculates the percentage of utilization for CPU and
memory, and if certain thresholds are exceeded then highlights the offending
line with an appropriate color.

The thresholds used are:
 * \>= 90% - high utilization   - RGBA(165,  29,  45, 77)
 * \>= 80% - medium utilization - RGBA(230,  97,   0, 77)                                                                                                    
 * \>= 70% - low utilization    - RGBA(246, 211,  45, 77)
 *   < 70% - no highlight