Refactor how CPU cache values are calculated
This new implementation takes into account the `shared_cpu_list` entry in each
`index#` directory from `/sys/devices/system/node/node#/cpu#`.

This means that once a cache value is extracted from `size`, for the
appropriate `level`, the `shared_cpu_list` is parsed and all sibling CPUs are
marked so that for this particular cache type and level it won't be added to
the total again.