Implement special handling, in the Apps page, for web browsers
Web browsers (installed as native packages), in the `.desktop` file, declare
their `Exec=` entries with binaries found in `PATH`. These executables are,
usually, just shims, links or similar for the real executables that lie in
`/usr/lib`, or `/opt`.

Another oddity is that the executables found in these alternative locations
are usually not named as their counterparts in `PATH`.

The "fix" implemented here essentially adds these extra paths to the binary
search path and performs any needed touch-ups to the executable names, as well
as allowing for multiple potential binary names in the `App` struct, to
compensate for the fact that multiple executables are valid for one
application.