Refactor proxy completely into an executable that will run as long as the main app. This process now known as gatherer and it communicates via Unix Domain Sockets and Shared Memory.
Sockets are used to send queries from the main application to the gatherer. The gatherer also uses this communication channel to send acknowledgement messages as well as signal that an operation has completed to the main application.

All data that is collected by the gatherer is stored in shared memory and is directly accessible by the main application without the need to (de)serialize data.

This infrastructure should provide the means to implement process termination, more complete GPU statistics, Intel GPU support, easier porting to other platforms, etc.