app_page: Move context menu handling to item
Gives the item responsibility for the context menu, allowing
us to remove the code from the page. We no longer need to do
anything like passing the widget as a pointer or computing the
points, so our logic is much simpler and safer.

This will also allow us to add touchscreen support if we add
an extra long press gesture on the item.