drive: Make the sidebar, in the performance page, a separate widget
Refactor the sidebar for drives, in the performance page, so that it is a
proper stand-alone widget. This is done to reduce complexity and ease the use
of bindings to control visibility of certain elements at runtime.

Also add support for a drive's serial number, WWN (World Wide Name) and
rotation rate, if applicable.