cpu: Add CPU power draw
read CPU power draw from `/sys/class/powercap/intel-rapl*/energy_uj`. For
security reasons, this file is unreadable as a regular user, so permissions must
be changed to read the data.

These security issues should be of no concern for the regular user, it's more
of a problem for servers.

Users are expected to manually change the file to be readable due to these
security issues.

In order to make the required file(s) readable a user can run the following:
```
sudo chmod a+r /sys/class/powercap/intel-rapl*/energy_uj 
```